削除機能が動かなくなっていたので、調整しました
