+++ templates/uploadAndDelete.html
@@ -58,9 +58,9 @@
         <section class="mt-5" id="send-to-delete-picture" style="padding-left: 10%; padding-right: 10%;">
 
             
-            <form class="row" method="post" action="/delete" id="delete-picture">
+            <form class="row" method="post" action="/delete">
                 <h1 class="text-center col-12">画像の削除</h1>
-                <div class="col-md-10">
+                <div class="col-md-10" id="delete-picture">
                     <select class="form-select form-select-lg" name="picture-list" id="select-picture">
                         <option id="selections">削除する画像を選択</option>
                     </select>
